requirements.txt
streamlit
python-dotenv
PyPDF2
sentence-transformers

# LangChain Core & Main
langchain
langchain-community
langchain-core
langchain-text-splitters

# Integrations
langchain-groq
langchain-huggingface
langchain-pinecone
pinecone-client
